Update according to refactored uuid frameworking from ArimAPI
